+++ src/main/java/app/dearobjet/backend/domain/payment/toss/TossPaymentsProperties.java
@@ -3,9 +3,11 @@ package app.dearobjet.backend.domain.payment.toss;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
 
 @Getter
 @Setter
+@Component
 @ConfigurationProperties(prefix = "toss")
 public class TossPaymentsProperties {
     private String secretKey;
